Remove `BaseTextArtifact`, revert `CsvRowArtifact` to subclass `TextArtifact`.
